refactor: add pay adaptor
